rename smh-organization to smh-app in repo
