+++ README.rst
@@ -1,4 +1,4 @@
-SMH Organization
+SMH App
 ------------------------
 
 - Python 3.7.x
@@ -20,12 +20,12 @@ Getting Started
 
 First clone the repository from Github and switch to the new directory::
 
-    $ git clone git@github.com:TransparentHealth/smh-organization
-    $ cd smh-organization
+    $ git clone git@github.com:TransparentHealth/smh-app
+    $ cd smh-app
 
 To setup your local environment you can use the quickstart make target `setup`, which will
-install Python (via pip) into a virtualenv named "smh_organization",
-and create a database via Postgres named "smh_organization" with all migrations run::
+install Python (via pip) into a virtualenv named "smh_app",
+and create a database via Postgres named "smh_app" with all migrations run::
 
     $ make setup
-    $ workon smh_organization
+    $ workon smh_app
