+++ smh_app/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for smh_organization project.
+Django settings for smh_app project.
 
 Generated by 'django-admin startproject' using Django 2.1.7.
 
@@ -54,12 +54,12 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'smh_organization.urls'
+ROOT_URLCONF = 'smh_app.urls'
 
 TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
-        'DIRS': [os.path.join(BASE_DIR, 'templates', 'smh_organization'), ],
+        'DIRS': [os.path.join(BASE_DIR, 'templates', 'smh_app'), ],
         'APP_DIRS': True,
         'OPTIONS': {
             'context_processors': [
@@ -74,7 +74,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'smh_organization.wsgi.application'
+WSGI_APPLICATION = 'smh_app.wsgi.application'
 
 
 # Database
@@ -83,7 +83,7 @@ WSGI_APPLICATION = 'smh_organization.wsgi.application'
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.postgresql',
-        'NAME': 'smh_organization',
+        'NAME': 'smh_app',
         'USER': '',
         'PASSWORD': '',
         'HOST': '',
