+++ smh_app/urls.py
@@ -1,4 +1,4 @@
-"""smh_organization URL Configuration
+"""smh_app URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/2.1/topics/http/urls/
