+++ smh_app/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for smh_organization project.
+WSGI config for smh_app project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'smh_organization.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'smh_app.settings')
 
 application = get_wsgi_application()
