+++ templates/smh_app/resources.html
